requirements.txt
autoflake==1.4
black==22.3.0
boto3==1.24.81
botocore==1.27.81
celery==5.2.7
DateTimeRange==1.2.0
decorator==5.1.1
Django==3.2.4
django-3-jet==1.0.8
django-admin-rangefilter==0.9.0
django-allauth==0.44.0
django-celery-beat==2.2.1
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-countries==7.3.2
django-crontab==0.7.1
django-extensions==3.2.1
django-filter==22.1
django-guardian==2.4.0
django-import-export==2.9.0
django-js-asset==2.0.0
django-localflavor==3.1
django-mailjet==0.3.1
django-multiselectfield==0.1.12
django-nested-admin==4.0.0
django-phonenumber-field==7.0.0
django-rest-auth==0.9.5
django-rest-swagger==2.1.2
django-reverse-admin==2.9.6
django-ses==3.1.2
django-shortuuidfield==0.1.3
django-simple-history==3.0.0
django-storages==1.13.1
django-taggit==3.0.0
djangorestframework==3.14.0
drf-nested-routers==0.93.3
drf-yasg==1.20.0
indian-cities==1.0.0
ipython==8.5.0
isort==5.10.1
numpy==1.22.2
pandas==1.4.0
Pillow==9.2.0
phonenumbers==8.12.43
pyarrow==8.0.0
pyotp==2.7.0
pytz==2022.2.1
python-dateutil==2.8.2
redis==3.5.3
reportlab==3.6.11
requests==2.28.1
requests-oauthlib==1.3.1
shortuuid==1.0.9
six==1.16.0
SQLAlchemy==1.4.31
